Split declarations into multiple files
