+++ tailormap-components/projects/bridge/typings/Components.d.ts
@@ -1,0 +1,17 @@
+import {
+  AppLayer,
+  GeoService,
+  MapComponent,
+} from './Mapcomponents';
+
+declare interface ViewerController {
+  mapComponent: MapComponent;
+
+  isDebug: () => boolean;
+
+  getService: (serviceId: number) => GeoService;
+  getAppLayerById: (appLayerId: number) => AppLayer;
+  getAppLayer: (serviceId: number, layerName: string) => AppLayer;
+
+  getVisibleLayers: (castToStrings?: boolean) => number[] | string[];
+}
